requirements.txt
boto3==1.35.99
botocore==1.35.99
fastapi==0.115.6
pyodbc==5.2.0
sqlmodel==0.0.22
uvicorn==0.34.0
python-multipart==0.0.20
pycryptodome==3.21.0
cryptography==44.0.0
infisicalsdk==1.0.3
BetterPassphrase==0.1.5
